Refactor steps.py into multiple files
